+++ src/cli/configuration_builder.ts
@@ -58,8 +58,6 @@ export default class ConfigurationBuilder {
 
   constructor({ argv, cwd }: INewConfigurationBuilderOptions) {
     this.cwd = cwd
-
-    ArgvParser.lint(argv)
     const parsedArgv = ArgvParser.parse(argv)
     this.args = parsedArgv.args
     this.options = parsedArgv.options